Add region selector to the GUI
Navigating to regions in the GUI has required typing in the region ID. To ease region identification, add a combo box with regions for the user to select by name. Populate the regions box with names of regions truncated to fit into the current minimum panel width and add abbreviations to ensure that names are unique. Since the list is often very long, use the TraitsUI `EnumEditor` with auto-compete support to find regions by typing in the start of their names.

Once the user selects a region, the GUI will immediately navigate to it. With the region ID box, the user could also enter additional regions separated by commas. To support this feature, add a check box labeled "Append" to append regions into the ID box without navigating to them. Once the user has appended all desired regions, unchecking the box will navigate to the aggregate region.