Expand registered images selectors to include modifiers just before the extension
Registered images are sometimes generated with additional modifiers before the extension, such as `annotationLevel5.mhd`. Include these files in the dropdowns by checking for existing file matches. Test file extensions directly rather than using the `read_sitk` check since it also counts the initial filename as a match, even if it does not have an extension in the list of supported 3D formats.