Position label annotations at the middle pixel
Label annotations have been placed at the label's centroid, but this position is not necessarily over the label itself, such as for labels that have an empty center or multiple components. To ensure that the annotation is over a label pixel, employ the same approach as for label navigation in the GUI, which finds the middle pixel in the list of pixel coordinates.

- Replace finding the centroid through region properties with `ontology.get_region_middle` to get the middle pixel
- Allow the `scaling` parameter in `ontology.get_region_middle` to be a non-NumPy sequence, and add a default value of None
- Turn off its debugging print statements