Provide option to position label annotation at centroid
Label edge images likely need annotations to be over the labels' centroids rather than a middle label pixel since these pixels are only on the outside of each label. Add a parameter to choose whether to place the annotation over the label itself (default), or at the label's centroid, which may not necessarily be a label pixel.