Add an optional dependencies group for `most` packages
The `all` package group contains almost all optional dependencies, many of which are necessary for typical operation, but some of which are rarely needed. As an in-between this `all` group and the base set of dependencies, add a `most` group that contains these mostly essential packages while omitting other packages such as AWS-related packages and Scikit-learn.