Move most of run script into module within `magmap` folder
The `run.py` script is not currently available during installations and more difficult to identify outside of the `magmap` namespace. Refactor most of this script into a separate module within the `magmap.io` subpackage. This changes fixes the inability to find the `run` module when running `magmap.gui.visualizer` from an installed MagellanMapper package.

- Move most of `run` into `magmap.io.load_env`
- Change `run` to be a wrapper entry point for this script
- Move the unhandled exception setting into `main` to avoid duplicating this call in the two scripts
- Duplicate the `freeze_support` call in the new script since it needs to be the first call in `if __name__ == '__main__'` (see https://docs.python.org/3/library/multiprocessing.html#multiprocessing.freeze_support), allowing this new script to similarly be launched directly