Add a Conda environment spec that includes `magellanmapper`
Install `magellanmapper` from Pip in a separate version of the main Conda environment specification so that MagellanMapper can be installed in Conda without downloading the repo. This specification will need to be manually synchronized with the main spec since there does not appear to be a way to load one spec from another.