Allow rearrangement of plot editors in the Atlas Editor
Add a `planes` attribute to specify which planes to show in the Atlas Editor. Continue to use the same default arrangement of one large vertical plot editors next to two smaller editors on top of one another, but expand their sizes to allow two plots side-by-side or above one another or one large plot depending on which planes are given.