Fix showing only one plot in the Atlas Editor
In the single plot layout, the plane can be set in any of the dropdowns, which currently leads to an error when this plane is selected in any dropdown but the first. Remove the need for indexing by filtering the planes for non-empty selections, removing None viewers, and converting max sizes to a dict accessed by the plane string.