Simplify rescaling and resizing to find the correct function based of argument
The wrapper function for rescaling and resizing has accepted separate arguments for each scenario, analogous to the underlying Scikit-image functions. As a simplification in this wrapper, merge the `rescale` into the `target_size` parameter by determining the function to call based on the size of the `target_size` parameter. Also, remove the anti-aliasing argument since it will default to True when downsampling in both scenarios.