Replace auto-rescaling with direct rescaling or resizing
The `resize_factor` atlas profile setting was meant to serve as a auto-scaler to approximate image shapes but has been rarely used since specific rescaling factors or target sizes are often required. As a simplification, remove this profile setting, including the "big" profile. Use the `target_size` parameter in the atlas transposer function to resize or rescale the image directly, without converting to a scaling factor to approximate a target shape.