Remove spacing/origin change in atlas transposer
Resizing/rescaling in the atlas transposer is typically to alter the physical dimensions of the image, where the number of pixels changes while the spacing remains the same. Also, the spacing should have been divided rather than multiplied if the physical dimensions were held constant. As a simplification, remove the spacing change to allow changing the physical dimensions of an atlas relative to a sample.