Workaround for file and directory import selection
The file browser for selecting files and/or directories is not working on some platforms and environments. As a workaround, replace the TraitsUI-based file browser with the one from Pyface, which works for selecting files or directories and setting a default path. This browser requires separate controls for selecting files or directories, so a separate directory browser button has been added. Add a preference for storing the last opened import path.