Fix `Protocol` error on Python < 3.8
The `Protocol` class for type hinting is only available in Python >= 3.8 or through the `typing_extensions` package. This package provides additional typing features for each version of Python and is already installed as sub-dependency of `bgatlas-api`. Given our move to incorporate type hinting more extensively in MagellanMapper, we will add this dependency explicitly.

- Make dependency on `typing_extensions` explicit
- Import `Protocol` from this package since it will import from `typing` if available (see https://stackoverflow.com/a/71995876/1911852 and https://github.com/python/typing_extensions/blob/main/src/typing_extensions.py#L457-L459)