Workaround to preserve original label identities for large values
Large label IDs may be lost when applying an Elastix Transformix filter. As a workaround suggested here (https://github.com/SuperElastix/SimpleElastix/issues/205#issuecomment-718852030), map the label identities to their indices in their unique array, which minimizes the chance of having large values. Default to turning off this feature to avoid applying it to non-label images, such as when repeating a transformation to multiple intensity images.