Set plot side for region label based on position in plot rather than data
The region label position changes orientation around the mouse pointer to maximize available space. The orientation has been based on data coordinates, which works when the full image is shown in the plot but not for zoomed plots, where the data midline usually differs from the plot midline. Change this midline to use the axes coordinates rather than the data coordinates.