Make full GUI dependencies optional
The fully integrated GUI uses a TraitsUI/Pyface/Mayavi stack with PyQt5 but is not necessary for running other parts of MM, including the individual Matplotlib-based GUI components. Move the full GUI dependencies into a separate, optional group, which allows MM to be installed as a library with a smaller dependency footprint. Include this GUI group in the `most` and `all` groups to ensure that it is still installed in the recommended pathways.