Fix leftover circle when exiting axes or changing planes with blitter
When the painting circle is managed by the blitter, the circle needs to remain in the figure after axes exit to blit away the artist. The circle object also needs to remain in the Plot Editor for removal from the blitter when drawing a new plane. Fix by removing the circle only when no blitter is present.