Update Plot Editor motion state earlier for repeated events
The motion handler has updated the stored state of positions and time at the end of the handler, but serial motion events appear to use outdated state information. While the impact appears to be minimal for panning, zooming exhibits a lag when quickly reversing direction from zooming in to out or vice versa, where the plot may continue to zoom in the current direction for a short while before reversing direction.

- As a fix, get the last state at the beginning of the motion handler and store the current state as soon as an event will be processed rather than skipped
- Re-update the data location after panning
- Change a print statement to debugging
- Fix type hint for `translate_coord`