Change shortcut for jumping to plane in ROI
The shortcut has been to right-click on the ROI plane, but control-left-click can emulate a right-click, which conflicts with the same shortcut used to add a blob. Change to "j"-left-click for "jump." Although this shortcut may not be as memorable, it is also less likely used since mouse scrolling and the arrow keys also control ROI plane scrolling. By preventing accidental jumping, the ROI is less likely to shift unexpectedly when viewing the next ROI.