Support simple spectral unmixing through linear channel subtraction
To reduce bleedthrough, add a setting to support subtracting one channel from another. Add an ROI profile setting, `spectral_unmixing`, that takes a dictionary specifying the channel to subtract from, the channel to subtract, and a multiplier for that channel's signal. Also, add type hints to getting an ROI profile.