Preprocess even undetected channels to allow for cross-channel processing
Spectral unmixing depends on channels other than the one being detected, but these channels are excluded from preprocessing and not available for the detector if they are not also set to be detected. To ensure that these channels are available, simply preprocess all channels. Although this leads to unnecessary preprocessing when spectral unmixing is not enabled, it simplifies blob detection rather than checking for this setting and allow for future tasks that employ other cross-channel processing.