Default to use the axes bounding box when fitting frame to image
Make providing the image shape optional when fitting a figure frame to its image by defaulting to use the bounding box of the first axes. This change allows fitting to an image rotated by a transformation that changes the image's bounding box but not the shape of the underlying image. Apply this change to extracting a stack of planes. Also, add a default value to the `aspect` parameter since it accepts None.