Store in the original intensity images when displaying images in the Plot Editor
The original intensity image may differ from the array of the plotted axes image, such as for additive blended images. These original images are required to set intensity ranges in the image adjustment panel, where changes to intensities are based on the original rather than the blended image. Add an attribute to the plotted image storage class to hold these images.