Workaround for intensity range synchronization slider jumping
The `RangeEditor` slider controls may shift unexpectedly, such as suddenly decreasing to 0 when shifting only slightly downward, or shifting upward before shifting downward. Clicking on a new position may also bring the slider there, but releasing the mouse button will cause the slider to pop back to its original positions. A shift in one intensity control can cause a further unexpected shift in the other control since we attempt to synchronize them. For example, the max intensity slider cannot be below the min slider, so an unintentional shift in the max slider to 0 will also cause the min slider to shift to 0. As a workaround, turn off this synchronization during intensity slider shifts. Leave it on when changing the auto setting to allow it to respond to directly toggling the auto control, and toggling it by moving a slider appears to respond to the initial shift rather than an unexpected subsequent shift.