Support ITK-Elastix for Transformix operations
Accept Transformix filters provided from ITK-Elastix in addition to SimpleITK. Detect the appropriate library to use based on the given filter type. Add a library method to support conversion from either SimpleITK or ITK Image objects to NumPy arrays and vice versa, and add support for replacing the array in an ITK Image.