Support registering ITK Image types
Accept ITK Image instances in `register_duo`, which also serves as a flag to use ITK-Elastix for registration. Continue to accept SimpleITK Image instances, using SimpleITK for registration if Elastix is enabled or converting the images to ITK Image types. Also, allow images given to the SimpleITK and ITK conversion functions to be None.