Remove obsolete transformation parameters update
The translation parameters have been extracted for overlaying single planes, but this overlay has been deprecated and commented out. Group and comment out this extraction with the overlay. This change will no longer save the translation parameters, but they are only a subset of the transformation, and the full transformation is still saved in the working directory.