Make ITK dependencies optional
When ITK-based dependencies were added as an alternative for SimpleITK, both libraries were made to be optional for core functionality. Move the these ITK-based dependencies into an `itk` install group, which will also be included in the `most` and `all` groups. Also, make the SimpleITK group constant's name consistent with the other extra installs constants.