Fix displaying atlas labels on rotated images
PR #471 introduced a regression where the transformation data class was added as an argument to the Matplotlib text annotation when only the transform itself should be added. Add this transform only when both the data class instance and transform are present since a transform of None also appears to override any default transform.