Allow using full intensity range in adjustment sliders
The adjustment sliders have attempted to adapt to the majority of signal in the image to increase precision of slider adjustments by limiting the range. The full range is sometimes necessary, however, to obtain the optimal viewing intensity. Change the sliders to the TraitsUI large range mode, which adds additional button to navigate to smaller or larger ranges. Remove the automated limit adjustments since the full range can now be navigated. The auto-intensity still attempts to select an appropriate starting intensity, which also positions the slider in an appropriate starting slider range. Additionally, reduce the number of decimal points in slider values and size of names since the large range sliders take up considerably more horizontal space.