Fix restoring intensity settings when some channels are deselected
The image overlayer expects all channel settings to be given in absolute rather than relative position, but the PlotEditor has supplied them relative to only the images shown. If a preceding channel was deselected, prior channel-specific settings would be given at the incorrect index and either applied to the wrong channel or ignored if the channel was not shown. Fix by converting these settings to absolute channel order for the overlayer. Update the overlayer docs to clarify the expected channel order.