Synchronize image adjustment controls after redrawing editor
Redrawing the editor has not updated the adjustment controls, leaving them in a stale state until the adjustment channel control is changed. This stale state has been used now to restore adjustment settings to the new image, but the settings need to be updated when the settings are not restored, fixed here.