Dynamically refresh the image on channel changes
To speed up responsiveness, automically refresh the image when the user changes the image channel control. This change requires using an ignore flag to ignore channel changes during the control's initialization. The refresh also serves as a way to fully refresh the image without reloading the image now that the Redraw button leaves some settings intact.