Decrease intensity high range for low values
The range adjustments in the `RangeEditor` `xslider` do not adapt below a high of 10, making slider adjustments difficult for low ranges, such as 0-1. For these values, use the image's near max, especially now that images loaded as registered images calculate this value. If this range is too small, the user can enter higher values manually into the control's text field. To allow the user to use the scroller afterward, expand the intensity high range to this manually input value. Although the range cannot be decreased, the control can be reset by refreshing the image (now by changing a channel).