Workaround Mayavi with TraitsUI 8
Mayavi gives an error when running with TraitsUI 8 (see https://github.com/enthought/mayavi/issues/1260). A workaround has been to set the `ETS_TOOLKIT="qt4"` environment variable, which worked on testing in macOS but not in Windows. As another workaround, set the `ETSConfig.toolkit` configuration instead, which avoids needing to set an environment variable and also works for Windows.