Remove direct reference to `itk` in setup
Specifying `itk` causes installations on macOS arm64 platforms to be unable to load `itk` for some reason. Remove this reference since `itk-elastix` will install `itk` anyway, and `itk` appears to work properly with this installation. `itk-elastix` also currently installs a later, pre-release version of `itk`. Note that the current release of `itk` (v5.3.0) does work on macOS arm64 if installed before the rest of the dependencies, for unclear reason.