Turn off on-the-fly blob detection in surrounding areas if blobs are given
Loading a saved ROI has also performed on-the-fly blob detection in the surrounding region to fill in these areas in case a border for the ROI is shown. It may not be intuitive to the user whether these border blobs are from the original ROI or newly detected blobs. For clarity and simplicity, turn off this detection.