Fix region export RGB values
Region export has generated colormaps for the number of labels in the image, but displayed images use a discrete colormap generated for a these labels plus a duplicated set of corresponding negative value labels. The colormap randomization requires this exact number of labels to generate an identical set of colors, meaning that the exported labels have had a different set of colors. Fix by using the same colormap generation wrapper function as used in the setup for labels image display.